Fix showing not required fields after qa item add
After adding item with quick add fields with contraints are shown even in
case constraints are not satisfied. Reason - after adding item fields are
cleared, but 'change' event is not fired (and even fired, we was catching
it incorrectly).

Fixed by correctly subscribe on 'change' event (jQuery requires space as
events separator [http://api.jquery.com/on/, see 'events']) and
triggering 'change' on TP side when fields are cleared.